feat!: publicly organize effects into submodules
