+++ src/effects/algebra.rs
@@ -1,4 +1,18 @@
-//! Implements [`Effect`] for tuples of effects up to size 8.
+//! [`Effect`] implementations for generic algebraic data types.
+//!
+//! The implementations here are on these foreign types...
+//! - tuples of `Effect`s
+//! - `Either<Left, Right>` where the `Left` and `Right` are both effects
+//! - the unit type `()` (trivial no-op effect)
+//!
+//! It's also worth noting that `Option<T>` where `T: Effect` also implements [`Effect`], but that is
+//! implemented in the [`iter`] module.
+//!
+//! Also relevant, [`Effect`] [can be derived] for any custom `struct`/`enum` where the internal
+//! types implement [`Effect`].
+//!
+//! [`iter`]: crate::effects::iter
+//! [can be derived]: crate::effect::Effect#derive
 
 use bevy::ecs::system::SystemParam;
 use bevy::prelude::*;
@@ -50,7 +64,7 @@ mod tests {
 
     use super::*;
     use crate::effects::number_data::NumberResource;
-    use crate::effects::res_set;
+    use crate::effects::resource::res_set;
     use crate::prelude::affect;
 
     proptest! {
